fed-analysis-server-inference/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1
concrete-ml==1.7.0
pandas